introduce VamComm to encapsulate VAm communication
also added some doc strings I forgot
also some clean up
